issue #58 move Employee class to solution file and use reduce functions
